PdftoPodcastApp/requirements.txt
streamlit
gtts
PyMuPDF
python-dotenv
google-generativeai